+++ classes/druid.lua
@@ -81,15 +81,20 @@ local function useNaturesGrace()
 end
 
 local function useFuryOfStormrage()
+    local buff = SAO.IsSoD() and 414800 or 81093; -- Fury of Stormrage (buff)
+    local talent = SAO.IsSoD() and 414799 or 17104; -- Fury of Stormrage (rune) or Fury of Stormrage (talent)
     SAO:CreateEffect(
         "natures_grace",
-        SAO.SOD,
-        414800, -- Fury of Stormrage (buff)
+        SAO.SOD + SAO.CATA,
+        buff,
         "aura",
         {
-            talent = 414799, -- Fury of Stormrage (rune)
+            talent = talent,
             overlay = { texture = "fury_of_stormrage", position = "Top" },
-            button = healingTouch,
+            buttons = {
+                [SAO.SOD] = healingTouch,
+                [SAO.CATA] = starfire,
+            },
         }
     );
 end